requirements.txt
fastapi == 0.115.12
uvicorn == 0.34.2
sqlalchemy == 2.0.40
python-dotenv == 1.1.0
asyncpg == 0.30.0